requirements.txt
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-core>=0.1.5
langgraph>=0.0.20
sentence-transformers>=2.2.2
chromadb>=0.4.22
transformers>=4.34.0
torch>=2.0.0
pypdf>=3.17.0
streamlit>=1.29.0
openai>=1.0.0  # OpenRouter uses OpenAI's API format
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
